Reworking the get-server.sh to install as non root.
Lots of folks get antsy when curl | bash'ing. To the extent we can
not use root thats great. So if a user has $HOME/bin or $HOME/.local/bin
on their PATH we can use those. Otherwise we can use /usr/local/bin or
/opt/bin. But the user needs to use us with sudo explicitly. Rather then us
escalating to it.

Also made the code more idomatic bash.

The docs will also need to be updated:

  curl -sf -L https://download.daytona.io/daytona/get-server.sh | sudo bash