Move to the DCO v1.1. (#189)
We were using a contributors file to track acceptance of the Apache
license by contributors. It was adding a lot of friction for first time
contributors. So we're going to move to the DCO v1.1 which only requires
that contributors sign or signoff on their commits.

For the contributions up until now, we will leave the contributors file
intact. From this commit forward, we will enforce the DCO and all
commits will be signed.

Signed-off-by: Chad Metcalf <metcalfc@gmail.com>